+++ src/pkg/sources/new.go
@@ -25,7 +25,6 @@ func New(pkgLocation string, pkgName string, opts zarfTypes.ZarfPackageOptions, 
 			PkgManifestSHA: sha,
 			TmpDir:         opts.PackageSource,
 			BundleLocation: pkgLocation,
-			Dev:            config.Dev,
 		}
 	} else {
 		platform := ocispec.Platform{
